# openapi/public.yaml
openapi: 3.1.0
info:
  title: API Kurir
  version: 1.0.0
  license:
    name: Proprietary
    identifier: LicenseRef-Proprietary
  description: >
    API ongkir internal Emisell. Endpoint customer pada /api/v1 kompatibel
    dengan kontrak shipping-cost RajaOngkir V2 dan memakai ID numerik dump
    region-service RajaOngkir; /v1 memakai kontrak canonical untuk client
    internal. Header key dan Bearer hanya cara autentikasi dan tidak mengubah
    bentuk respons. Uang menggunakan integer rupiah dan berat menggunakan gram.
servers:
  - url: /
    description: Host API Kurir aktif
security:
  - bearerAuth: []
  - rajaOngkirKeyAuth: []
...
paths:
...
  /api/v1/destination/domestic-destination:
    get:
      tags: [RajaOngkir V2 - Locations]
      operationId: searchDomesticDestination
      summary: Mencari master lokasi lokal dengan respons RajaOngkir V2
      security:
        - bearerAuth: []
        - rajaOngkirKeyAuth: []
      parameters:
        - name: search
          in: query
          required: true
          description: >
            Nama wilayah, kode pos, atau alamat mentah. Awalan administratif,
            tanda baca, nomor non-kode-pos, dan token duplikat dinormalisasi.
          schema:
            type: string
            minLength: 2
          example: Beji Depok
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            maximum: 100000
            default: 0
      responses:
        "200":
          description: Hasil pencarian
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RajaOngkirDestinationSearchResponse"
        "400":
          description: Parameter pencarian tidak valid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RajaOngkirErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "422":
          description: Parameter pencarian tidak lengkap atau tidak valid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RajaOngkirErrorResponse"
  /api/v1/destination/province:
    get:
      tags: [RajaOngkir V2 - Locations]
      operationId: listProvinces
      summary: Mendapatkan daftar provinsi lokal
      responses:
        "200":
          description: Daftar provinsi
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LocationHierarchyResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/destination/city/{province_id}:
    get:
      tags: [RajaOngkir V2 - Locations]
      operationId: listCitiesByProvince
      summary: Mendapatkan kota/kabupaten berdasarkan ID provinsi
      parameters:
        - $ref: "#/components/parameters/ProvinceID"
      responses:
        "200":
          description: Daftar kota/kabupaten
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LocationHierarchyResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/destination/district/{city_id}:
    get:
      tags: [RajaOngkir V2 - Locations]
      operationId: listDistrictsByCity
      summary: Mendapatkan kecamatan berdasarkan ID kota/kabupaten
      parameters:
        - $ref: "#/components/parameters/CityID"
      responses:
        "200":
          description: Daftar kecamatan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LocationHierarchyResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/destination/sub-district/{district_id}:
    get:
      tags: [RajaOngkir V2 - Locations]
      operationId: listSubdistrictsByDistrict
      summary: Mendapatkan kelurahan/desa berdasarkan ID kecamatan
      parameters:
        - $ref: "#/components/parameters/DistrictID"
      responses:
        "200":
          description: Daftar kelurahan/desa
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LocationHierarchyResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/couriers:
    get:
      tags: [RajaOngkir V2 - Couriers]
      operationId: listCouriers
      summary: Mendapatkan kurir dan layanan aktif
      responses:
        "200":
          description: Daftar kurir
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CourierListResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/v1/calculate/domestic-cost:
    post:
      tags: [RajaOngkir V2 - Rates]
      operationId: calculateDomesticCost
      summary: Menghitung ongkir dari rate card lokal atau exact provider quote
      security:
        - bearerAuth: []
        - rajaOngkirKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/MerchantIDHeader"
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/RajaOngkirCalculateCostRequest"
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/RajaOngkirCalculateCostRequest"
      responses:
        "200":
          description: Pilihan layanan dan biaya
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RajaOngkirCalculateCostResponse"
        "400":
          description: Parameter atau lokasi tidak valid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RajaOngkirErrorResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          description: Merchant belum mengaktifkan provider pengiriman
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RajaOngkirErrorResponse"
        "415":
          $ref: "#/components/responses/RajaOngkirUnsupportedMediaType"
        "422":
          description: Courier tidak valid atau tarif belum tersedia
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RajaOngkirErrorResponse"
        "502":
          description: Provider gagal atau menolak credential
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RajaOngkirErrorResponse"
        "503":
          description: Kuota provider habis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RajaOngkirErrorResponse"
  /api/v1/calculate/district/domestic-cost:
    post:
      tags: [RajaOngkir V2 - Rates]
      operationId: calculateDistrictDomesticCost
      summary: Menghitung ongkir berdasarkan ID kecamatan seperti RajaOngkir V2
      description: >
        Pada public API tanpa X-Emisell-Merchant-ID, courier wajib dikirim seperti
        RajaOngkir V2. Pada Emisell Gateway dengan Merchant ID, courier tidak
        diperlukan dan nilai dari client lama diabaikan; API Kurir mengambil
        kurir dari service yang dipilih merchant. Provider bawaan Emisell memakai
        pool credential platform secara internal, sedangkan provider RajaOngkir
        BYOK memakai credential tenant; credential_id tidak dikirim Main Service.
      security:
        - bearerAuth: []
        - rajaOngkirKeyAuth: []
      parameters:
        - $ref: "#/components/parameters/MerchantIDHeader"
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: "#/components/schemas/RajaOngkirCalculateCostRequest"
          multipart/form-data:
            schema:
              $ref: "#/components/schemas/RajaOngkirCalculateCostRequest"
      responses:
        "200":
          description: Pilihan layanan dan biaya dalam kontrak RajaOngkir V2
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RajaOngkirCalculateCostResponse"
        "400":
          description: Parameter atau lokasi tidak valid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RajaOngkirErrorResponse"
        "401":
          description: API key tidak valid
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RajaOngkirErrorResponse"
        "409":
          description: Merchant belum mengaktifkan provider pengiriman
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RajaOngkirErrorResponse"
        "415":
          $ref: "#/components/responses/RajaOngkirUnsupportedMediaType"
        "422":
          description: Courier public tidak valid atau merchant belum memilih service checkout
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RajaOngkirErrorResponse"
        "502":
          description: Provider gagal atau menolak credential
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RajaOngkirErrorResponse"
        "503":
          description: Kuota provider habis
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RajaOngkirErrorResponse"
...
  /v1/destination/domestic-destination:
    get:
      tags: [Canonical - Locations]
      operationId: searchCanonicalDomesticDestination
      summary: Mencari master lokasi dengan public ID canonical
      description: >
        Endpoint internal API Kurir. Alamat mentah dinormalisasi dan hasil
        memakai ID loc_idn_*. Jangan kirim ID numerik dari /api/v1.
      parameters:
        - name: search
          in: query
          required: true
          description: Nama wilayah, alamat mentah, atau kode pos
          schema:
            type: string
            minLength: 2
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 50
            default: 20
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            maximum: 100000
            default: 0
      responses:
        "200":
          description: Lokasi canonical yang cocok
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LocationSearchResponse"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/destination/province:
    get:
      tags: [Canonical - Locations]
      operationId: listCanonicalProvinces
      summary: Mendapatkan provinsi dengan public ID canonical
      responses:
        "200":
          description: Daftar provinsi canonical
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LocationHierarchyResponse"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/destination/city/{province_id}:
    get:
      tags: [Canonical - Locations]
      operationId: listCanonicalCitiesByProvince
      summary: Mendapatkan kota/kabupaten dengan public ID canonical
      parameters:
        - $ref: "#/components/parameters/ProvinceID"
      responses:
        "200":
          description: Daftar kota/kabupaten canonical
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LocationHierarchyResponse"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /v1/destination/district/{city_id}:
    get:
      tags: [Canonical - Locations]
      operationId: listCanonicalDistrictsByCity
      summary: Mendapatkan kecamatan dengan public ID canonical
      parameters:
        - $ref: "#/components/parameters/CityID"
      responses:
        "200":
          description: Daftar kecamatan canonical
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LocationHierarchyResponse"
        "400":
          $ref: "#/components/responses/InvalidRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API key
    rajaOngkirKeyAuth:
      type: apiKey
      in: header
      name: key
      description: Alternatif pengiriman customer API key; tidak mengubah kontrak path
    serviceBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Dedicated Emisell service API key
      description: >
        Key Main Service hasil generate dashboard dengan scope gateway:access,
        atau secret statis API_KEYS sebagai recovery.
    serviceKeyAuth:
      type: apiKey
      in: header
      name: key
      description: >
        Key Main Service hasil generate dashboard dengan scope gateway:access,
        atau secret statis API_KEYS sebagai recovery. Hanya disimpan pada backend Emisell.
    merchantContextAuth:
      type: apiKey
      in: header
      name: X-Emisell-Merchant-ID
      description: >
        Merchant ID stabil yang ditentukan backend Emisell. Wajib untuk semua
        endpoint /api/v1/integrations dan tidak diterima dari browser.
    adminBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Admin API key
    partnerBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: Emisell partner access key
      description: Access key berformat epk_live_* yang terikat ke satu provider.
...